fix(request): made error subscription mandatory for a request
Global io error events not firing for requests that do not make it to DeployR.
By making error subscription mandatory internally, all errors encountered will
fire both a transaction level error and a global error. It is still up to the
user to subscribe to both if they want to be notified.

Closes #2